fix(pairing): preserve doctor and SDK contracts
