+++ docs/.generated/plugin-sdk-api-baseline.sha256
@@ -1,2 +1,2 @@
-343feb0c0d803597d1ffe3771f6e7e214e3a5036ab630c71a7646b3dd80b61f2  plugin-sdk-api-baseline.json
-a7f20566d8e77fe5105b510e2631b0dff3c5b0a09daf3cd8568e18f3f56dca49  plugin-sdk-api-baseline.jsonl
+75edbe496d57c19d3ea53e51908c6c7aa37ed8a5c7311234c7ce17eea0775db8  plugin-sdk-api-baseline.json
+f21d330f28d93248bd6a7d96780452d6a969f5d1fcca1b9246ceb5a68e589b9f  plugin-sdk-api-baseline.jsonl
