+++ src/commands/doctor-state-migrations.test.ts
@@ -9,7 +9,7 @@ import { afterEach, beforeAll, describe, expect, it, vi } from "vitest";
 import type { OpenClawConfig } from "../config/config.js";
 import type { SessionEntry } from "../config/sessions/types.js";
 import { requireNodeSqlite } from "../infra/node-sqlite.js";
-import { readChannelPairingStateSnapshot } from "../pairing/pairing-store.js";
+import { readChannelPairingStateSnapshot } from "../pairing/pairing-store-sqlite.js";
 import {
   createPluginStateKeyedStore,
   resetPluginStateStoreForTests,
