+++ src/plugin-sdk/channel-pairing.ts
@@ -9,7 +9,7 @@ export {
   readChannelAllowFromStore,
   readChannelAllowFromStoreSync,
 } from "../pairing/pairing-store.js";
-export { resolveChannelAllowFromPath } from "./channel-pairing-paths.js";
+export { resolveChannelAllowFromPath } from "../pairing/pairing-store.js";
 import { issuePairingChallenge } from "../pairing/pairing-challenge.js";
 import type { PluginRuntime } from "../plugins/runtime/types.js";
 import { createScopedPairingAccess } from "./pairing-access.js";
